+++ packages/@valapi/lib/src/__tests__/resource.ts
@@ -1,12 +1,12 @@
-import { CrosshairColor, ItemTypeId, Locale, QueueId, Region } from "../index";
+import { CrosshairColor, CrosshairHexColor, ItemTypeId, Locale, QueueId, Region } from "../index";
 
 describe("lib.resource", () => {
     test("resources", () => {
-        expect(CrosshairColor.fromColor("White")).toBe("0");
+        expect(CrosshairColor.fromName("White")).toBe("0");
         expect(CrosshairColor.fromID("1")).toBe("Green");
 
-        expect(CrosshairColor.fromColorHex("Pink")).toBe("FF00FF");
-        expect(CrosshairColor.fromIdHex("00FFFF")).toBe("Cyan");
+        expect(CrosshairHexColor.fromName("Pink")).toBe("FF00FF");
+        expect(CrosshairHexColor.fromHex("00FFFF")).toBe("Cyan");
 
         expect(ItemTypeId.fromName("Agents")).toBe("01bb38e1-da47-4e6a-9b3d-945fe4655707");
         expect(ItemTypeId.fromID("e7c63390-eda7-46e0-bb7a-a6abdacd2433")).toBe("Skins");