+++ packages/@valapi/lib/src/index.ts
@@ -6,6 +6,7 @@ export { ValError } from "./client/ValError";
 // resources
 
 export * as CrosshairColor from "./resources/CrosshairColor";
+export * as CrosshairHexColor from "./resources/CrosshairHexColor";
 export * as ItemTypeId from "./resources/ItemTypeId";
 export * as Locale from "./resources/Locale";
 export * as QueueId from "./resources/QueueId";