+++ packages/@valapi/lib/src/resources/QueueId.ts
@@ -1,3 +1,5 @@
+import { ValError } from "../client/ValError";
+
 export const Default = <const>{
     Unrated: "unrated",
     Competitive: "competitive",
@@ -16,14 +18,14 @@ export const Default = <const>{
 };
 
 export type Name = keyof typeof Default;
-export type ID = Exclude<(typeof Default)[Name], (typeof Default)["New_Map"]>;
+export type ID = Exclude<(typeof Default)[Name], (typeof Default)["New_Map"] | "">;
 
 /**
  * Change from {@link Name} to {@link ID}
  * @param {Name} x Name
  * @returns {ID} ID
  */
-export function fromName(x: Name): ID | undefined {
+export function fromName(x: Name): ID {
     return <ID>Default[x];
 }
 
@@ -32,12 +34,16 @@ export function fromName(x: Name): ID | undefined {
  * @param {ID} x ID
  * @returns {Name} Name
  */
-export function fromID(x: ID): Name | undefined {
+export function fromID(x: ID): Name {
     for (const data of Object.entries(Default)) {
         if (typeof data[1] == "string" && data[1] == x) {
             return <Name>data[0];
         }
     }
 
-    return undefined;
+    throw new ValError({
+        name: "Resource_Error",
+        message: "Resource Not Found",
+        data: x
+    });
 }