+++ packages/valorant.ts/src/index.ts
@@ -1,7 +1,7 @@
 // packages
 
 export { AuthClient, AuthCore, AuthService } from "@valapi/auth";
-export { ValError, ValRegion, CrosshairColor, ItemTypeId, Locale, QueueId, Region, ValEncryption } from "@valapi/lib";
+export { ValError, ValRegion, CrosshairColor, CrosshairHexColor, ItemTypeId, Locale, QueueId, Region, ValEncryption } from "@valapi/lib";
 export { RiotApi, RiotApiRegion, RiotApiService } from "@valapi/riot-api";
 export { ValorantApiCom, ValorantApiComService } from "@valapi/valorant-api.com";
 export { WebClient, WebClientRegion, WebClientService } from "@valapi/web-client";